feat:Modularize gluejob Separated gluejobs into two parts: csv2parquet.py: to convert csv from staging bucket to parquet in transformed bucket ETL.py: for data processing Resolve DE-41
